Rename JSON array helper CIRC-381
