+++ src/main/java/org/folio/circulation/domain/notice/TemplateContextUtil.java
@@ -18,7 +18,7 @@ import org.folio.circulation.domain.RequestType;
 import org.folio.circulation.domain.ServicePoint;
 import org.folio.circulation.domain.User;
 import org.folio.circulation.domain.policy.LoanPolicy;
-import org.folio.circulation.support.json.JsonArrayHelper;
+import org.folio.circulation.support.json.JsonObjectArrayPropertyFetcher;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 
@@ -146,7 +146,7 @@ public class TemplateContextUtil {
   }
 
   private static JsonObject createItemContext(Item item) {
-    String contributorNamesToken = JsonArrayHelper.toStream(item.getContributorNames())
+    String contributorNamesToken = JsonObjectArrayPropertyFetcher.toStream(item.getContributorNames())
       .map(o -> o.getString("name"))
       .collect(joining("; "));
