+++ src/test/java/api/loans/LoanAPITests.java
@@ -17,7 +17,7 @@ import static java.lang.String.format;
 import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
 import static org.folio.circulation.domain.representations.ItemProperties.CALL_NUMBER_COMPONENTS;
 import static org.folio.circulation.domain.representations.LoanProperties.BORROWER;
-import static org.folio.circulation.support.json.JsonArrayHelper.toList;
+import static org.folio.circulation.support.json.JsonObjectArrayPropertyFetcher.toList;
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
