+++ src/test/java/api/requests/HoldShelfClearanceReportTests.java
@@ -1,7 +1,7 @@
 package api.requests;
 
 import static java.net.HttpURLConnection.HTTP_OK;
-import static org.folio.circulation.support.json.JsonArrayHelper.toList;
+import static org.folio.circulation.support.json.JsonObjectArrayPropertyFetcher.toList;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
