+++ src/test/java/api/requests/PickSlipsTests.java
@@ -29,7 +29,7 @@ import org.folio.circulation.domain.Item;
 import org.folio.circulation.domain.ItemStatus;
 import org.folio.circulation.domain.Location;
 import org.folio.circulation.domain.User;
-import org.folio.circulation.support.json.JsonArrayHelper;
+import org.folio.circulation.support.json.JsonObjectArrayPropertyFetcher;
 import org.folio.circulation.support.http.client.IndividualResource;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
@@ -168,7 +168,7 @@ public class PickSlipsTests extends APITests {
       .withHoldingsRecord(itemResource.getHoldingsRecord().getJson())
       .withInstance(itemResource.getInstance().getJson());
 
-    String contributorNames = JsonArrayHelper.toStream(item.getContributorNames())
+    String contributorNames = JsonObjectArrayPropertyFetcher.toStream(item.getContributorNames())
       .map(this::getName)
       .collect(joining("; "));
 
@@ -435,7 +435,7 @@ public class PickSlipsTests extends APITests {
   }
 
   private Stream<JsonObject> getPickSlipsStream(Response response) {
-    return JsonArrayHelper.toStream(response.getJson(), PICK_SLIPS_KEY);
+    return JsonObjectArrayPropertyFetcher.toStream(response.getJson(), PICK_SLIPS_KEY);
   }
 
   private List<JsonObject> getPickSlipsList(Response response) {
